Improve API efficiency and handling of unmanaged nodes (#506)
* Retrieve richer node information from DescribeInstances

* Read ASG and Managed state directly from instance tags when available

* Improve handling of empty interruption events

* reword log message for getNodeInfo

* Avoid unnecessary ASG API calls if ASG tag propagation is on

* Fix configuration env var name for consistency

* Add unit tests, tidy log wording

* clean extra comments from new tests

* Add new config key to Print*Args functions

* Add mock ASG tag to localstack instances for E2E tests

* Update pkg/config/config.go

+++ pkg/monitor/sqsevent/sqs-monitor.go
@@ -33,6 +33,8 @@ import (
 const (
 	// SQSTerminateKind is a const to define an SQS termination kind of interruption event
 	SQSTerminateKind = "SQS_TERMINATE"
+	// ASGTagName is the name of the instance tag whose value is the AutoScaling group name
+	ASGTagName = "aws:autoscaling:groupName"
 )
 
 // ErrNodeStateNotRunning forwards condition that the instance is terminated thus metadata missing
@@ -40,14 +42,15 @@ var ErrNodeStateNotRunning = errors.New("node metadata unavailable")
 
 // SQSMonitor is a struct definition that knows how to process events from Amazon EventBridge
 type SQSMonitor struct {
-	InterruptionChan chan<- monitor.InterruptionEvent
-	CancelChan       chan<- monitor.InterruptionEvent
-	QueueURL         string
-	SQS              sqsiface.SQSAPI
-	ASG              autoscalingiface.AutoScalingAPI
-	EC2              ec2iface.EC2API
-	CheckIfManaged   bool
-	ManagedAsgTag    string
+	InterruptionChan        chan<- monitor.InterruptionEvent
+	CancelChan              chan<- monitor.InterruptionEvent
+	QueueURL                string
+	SQS                     sqsiface.SQSAPI
+	ASG                     autoscalingiface.AutoScalingAPI
+	EC2                     ec2iface.EC2API
+	CheckIfManaged          bool
+	AssumeAsgTagPropagation bool
+	ManagedAsgTag           string
 }
 
 // Kind denotes the kind of event that is processed
@@ -66,25 +69,44 @@ func (m SQSMonitor) Monitor() error {
 	failedEvents := 0
 	for _, message := range messages {
 		interruptionEvent, err := m.processSQSMessage(message)
+		dropMessage := false
 		switch {
 		case errors.Is(err, ErrNodeStateNotRunning):
 			// If the node is no longer running, just log and delete the message.  If message deletion fails, count it as an error.
 			log.Warn().Err(err).Msg("dropping event for an already terminated node")
-			errs := m.deleteMessages([]*sqs.Message{message})
-			if len(errs) > 0 {
-				log.Err(errs[0]).Msg("error deleting event for already terminated node")
-				failedEvents++
-			}
+			dropMessage = true
 
 		case err != nil:
 			// Log errors and record as failed events
 			log.Err(err).Msg("ignoring event due to error")
 			failedEvents++
 
-		case err == nil && interruptionEvent != nil && interruptionEvent.Kind == SQSTerminateKind:
+		case interruptionEvent == nil:
+			log.Debug().Msg("dropping non-actionable event")
+			dropMessage = true
+
+		case m.CheckIfManaged && !interruptionEvent.IsManaged:
+			// This event isn't for an instance that is managed by this process
+			log.Debug().Str("instance-id", interruptionEvent.InstanceID).Msg("dropping event for unmanaged node")
+			dropMessage = true
+
+		case interruptionEvent.Kind == SQSTerminateKind:
 			// Successfully processed SQS message into a SQSTerminateKind interruption event
 			log.Debug().Msgf("Sending %s interruption event to the interruption channel", SQSTerminateKind)
 			m.InterruptionChan <- *interruptionEvent
+
+		default:
+			eventJSON, _ := json.MarshalIndent(interruptionEvent, " ", "    ")
+			log.Warn().Msgf("dropping event of an unrecognized kind: %s", eventJSON)
+			dropMessage = true
+		}
+
+		if dropMessage {
+			errs := m.deleteMessages([]*sqs.Message{message})
+			if len(errs) > 0 {
+				log.Err(errs[0]).Msg("Error deleting message from SQS")
+				failedEvents++
+			}
 		}
 	}
 
@@ -103,45 +125,20 @@ func (m SQSMonitor) processSQSMessage(message *sqs.Message) (*monitor.Interrupti
 		return nil, err
 	}
 
-	interruptionEvent := monitor.InterruptionEvent{}
-
 	switch event.Source {
 	case "aws.autoscaling":
-		interruptionEvent, err = m.asgTerminationToInterruptionEvent(event, message)
-		if err != nil {
-			return nil, err
-		}
+		return m.asgTerminationToInterruptionEvent(event, message)
+
 	case "aws.ec2":
 		if event.DetailType == "EC2 Instance State-change Notification" {
-			interruptionEvent, err = m.ec2StateChangeToInterruptionEvent(event, message)
+			return m.ec2StateChangeToInterruptionEvent(event, message)
 		} else if event.DetailType == "EC2 Spot Instance Interruption Warning" {
-			interruptionEvent, err = m.spotITNTerminationToInterruptionEvent(event, message)
+			return m.spotITNTerminationToInterruptionEvent(event, message)
 		} else if event.DetailType == "EC2 Instance Rebalance Recommendation" {
-			interruptionEvent, err = m.rebalanceRecommendationToInterruptionEvent(event, message)
-		}
-		if err != nil {
-			return nil, err
-		}
-	default:
-		return nil, fmt.Errorf("Event source (%s) is not supported", event.Source)
-	}
-
-	// Bail if empty event is returned after parsing
-	if interruptionEvent.EventID == "" {
-		return nil, nil
-	}
-
-	if m.CheckIfManaged {
-		isManaged, err := m.isInstanceManaged(interruptionEvent.InstanceID)
-		if err != nil {
-			return &interruptionEvent, err
-		}
-		if !isManaged {
-			return nil, nil
+			return m.rebalanceRecommendationToInterruptionEvent(event, message)
 		}
 	}
-
-	return &interruptionEvent, err
+	return nil, fmt.Errorf("Event source (%s) is not supported", event.Source)
 }
 
 // receiveQueueMessages checks the configured SQS queue for new messages
@@ -182,8 +179,19 @@ func (m SQSMonitor) deleteMessages(messages []*sqs.Message) []error {
 	return errs
 }
 
-// retrieveNodeName queries the EC2 API to determine the private DNS name for the instanceID specified
-func (m SQSMonitor) retrieveNodeName(instanceID string) (string, error) {
+// NodeInfo is relevant information about a single node
+type NodeInfo struct {
+	AsgName    string
+	InstanceID string
+	IsManaged  bool
+	Name       string
+	Tags       map[string]string
+}
+
+// getNodeInfo returns the NodeInfo record for the given instanceID.
+//
+// The data is retrieved from the EC2 API.
+func (m SQSMonitor) getNodeInfo(instanceID string) (*NodeInfo, error) {
 	result, err := m.EC2.DescribeInstances(&ec2.DescribeInstancesInput{
 		InstanceIds: []*string{
 			aws.String(instanceID),
@@ -192,22 +200,20 @@ func (m SQSMonitor) retrieveNodeName(instanceID string) (string, error) {
 	if err != nil {
 		if aerr, ok := err.(awserr.Error); ok && aerr.Code() == "InvalidInstanceID.NotFound" {
 			log.Warn().Msgf("No instance found with instance-id %s", instanceID)
-			return "", ErrNodeStateNotRunning
+			return nil, ErrNodeStateNotRunning
 		}
-		return "", err
+		return nil, err
 	}
 	if len(result.Reservations) == 0 || len(result.Reservations[0].Instances) == 0 {
 		log.Warn().Msgf("No instance found with instance-id %s", instanceID)
-		return "", ErrNodeStateNotRunning
+		return nil, ErrNodeStateNotRunning
 	}
 
 	instance := result.Reservations[0].Instances[0]
-	nodeName := *instance.PrivateDnsName
-	log.Debug().Msgf("Got nodename from private ip %s", nodeName)
 	instanceJSON, _ := json.MarshalIndent(*instance, " ", "    ")
 	log.Debug().Msgf("Got instance data from ec2 describe call: %s", instanceJSON)
 
-	if nodeName == "" {
+	if *instance.PrivateDnsName == "" {
 		state := "unknown"
 		// safe access instance.State potentially being nil
 		if instance.State != nil {
@@ -215,28 +221,61 @@ func (m SQSMonitor) retrieveNodeName(instanceID string) (string, error) {
 		}
 		// anything except running might not contain PrivateDnsName
 		if state != ec2.InstanceStateNameRunning {
-			return "", fmt.Errorf("node: '%s' in state '%s': %w", instanceID, state, ErrNodeStateNotRunning)
+			return nil, fmt.Errorf("node: '%s' in state '%s': %w", instanceID, state, ErrNodeStateNotRunning)
 		}
-		return "", fmt.Errorf("unable to retrieve PrivateDnsName name for '%s' in state '%s'", instanceID, state)
+		return nil, fmt.Errorf("unable to retrieve PrivateDnsName name for '%s' in state '%s'", instanceID, state)
 	}
-	return nodeName, nil
+
+	nodeInfo := &NodeInfo{
+		Name:       *instance.PrivateDnsName,
+		InstanceID: instanceID,
+		Tags:       make(map[string]string),
+		IsManaged:  true,
+	}
+	for _, t := range (*instance).Tags {
+		nodeInfo.Tags[*t.Key] = *t.Value
+		if *t.Key == ASGTagName {
+			nodeInfo.AsgName = *t.Value
+		}
+	}
+
+	if nodeInfo.AsgName == "" && !m.AssumeAsgTagPropagation {
+		// If ASG tags are not propagated we might need to use the API
+		// to retrieve the ASG name
+		nodeInfo.AsgName, err = m.retrieveAutoScalingGroupName(nodeInfo.InstanceID)
+		if err != nil {
+			return nil, fmt.Errorf("unable to retrieve AutoScaling group: %w", err)
+		}
+	}
+
+	if m.CheckIfManaged && nodeInfo.Tags[m.ManagedAsgTag] == "" {
+		if m.AssumeAsgTagPropagation {
+			nodeInfo.IsManaged = false
+		} else {
+			// if ASG tags are not propagated we might have to check the ASG directly
+			nodeInfo.IsManaged, err = m.isASGManaged(nodeInfo.AsgName, nodeInfo.InstanceID)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+	infoJSON, _ := json.MarshalIndent(nodeInfo, " ", "    ")
+	log.Debug().Msgf("Got node info from AWS: %s", infoJSON)
+
+	return nodeInfo, nil
 }
 
-// isInstanceManaged returns whether the instance specified should be managed by node termination handler
-func (m SQSMonitor) isInstanceManaged(instanceID string) (bool, error) {
-	if instanceID == "" {
-		return false, fmt.Errorf("Instance ID was empty when calling isInstanceManaged")
-	}
-	asgName, err := m.retrieveAutoScalingGroupName(instanceID)
+// isASGManaged returns whether the autoscaling group should be managed by node termination handler
+func (m SQSMonitor) isASGManaged(asgName string, instanceID string) (bool, error) {
 	if asgName == "" {
-		return false, err
+		return false, nil
 	}
 	asgFilter := autoscaling.Filter{Name: aws.String("auto-scaling-group"), Values: []*string{aws.String(asgName)}}
 	asgDescribeTagsInput := autoscaling.DescribeTagsInput{
 		Filters: []*autoscaling.Filter{&asgFilter},
 	}
 	isManaged := false
-	err = m.ASG.DescribeTagsPages(&asgDescribeTagsInput, func(resp *autoscaling.DescribeTagsOutput, next bool) bool {
+	err := m.ASG.DescribeTagsPages(&asgDescribeTagsInput, func(resp *autoscaling.DescribeTagsOutput, next bool) bool {
 		for _, tag := range resp.Tags {
 			if *tag.Key == m.ManagedAsgTag {
 				isManaged = true
@@ -248,11 +287,11 @@ func (m SQSMonitor) isInstanceManaged(instanceID string) (bool, error) {
 		return true
 	})
 
-	if !isManaged {
-		log.Debug().
-			Str("instance_id", instanceID).
-			Msgf("The instance's Auto Scaling Group is not tagged as managed with tag key: %s", m.ManagedAsgTag)
-	}
+	log.Debug().
+		Str("instance_id", instanceID).
+		Str("tag_key", m.ManagedAsgTag).
+		Bool("is_managed", isManaged).
+		Msg("directly checked if instance's Auto Scaling Group is managed")
 	return isManaged, err
 }
 
@@ -271,5 +310,9 @@ func (m SQSMonitor) retrieveAutoScalingGroupName(instanceID string) (string, err
 		return "", nil
 	}
 	asgName := asgs.AutoScalingInstances[0].AutoScalingGroupName
-	return *asgName, err
+	log.Debug().
+		Str("instance_id", instanceID).
+		Str("asg_name", *asgName).
+		Msg("performed API lookup of instance ASG")
+	return *asgName, nil
 }
